backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic>=2.0.0
pydantic-settings
sqlalchemy>=2.0.0
psycopg2-binary
asyncpg
redis
aiokafka
minio
pymilvus
requests
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-milvus
sentence-transformers
unstructured
